+++ Configuration/Options/AuthOptions.cs
@@ -1,0 +1,8 @@
+namespace Navislamia.Configuration.Options;
+
+public class AuthOptions
+{
+    public string Ip { get; set; }
+    public ushort Port { get; set; }
+    public int Index { get; set; }
+}
