+++ Configuration/Options/GameOptions.cs
@@ -1,0 +1,8 @@
+namespace Navislamia.Configuration.Options;
+
+public class GameOptions
+{
+    public string Ip { get; set; }
+    public ushort Port { get; set; }
+
+}
