+++ Configuration/Options/UploadOptions.cs
@@ -1,0 +1,7 @@
+namespace Navislamia.Configuration.Options;
+
+public class UploadOptions
+{
+    public string Ip { get; set; }
+    public ushort Port { get; set; }
+}
