+++ DevConsole/appsettings.Dev.json
@@ -1,0 +1,6 @@
+{
+  "Server": {
+    "name": "Navislamia - DEV",
+    "index": 80
+  }
+}
