+++ DevConsole/appsettings.Prod.json
@@ -1,0 +1,6 @@
+{
+  "Server": {
+    "name": "Navislamia - PROD",
+    "index": 90
+  }
+}
