+++ DevConsole/appsettings.json
@@ -1,7 +1,7 @@
 {
   "Server": {
     "name": "Navislamia",
-    "index": 0,
+    "index": 1,
     "screenshotUrl": "about:new",
     "isAdultServer": false
   },
@@ -15,16 +15,22 @@
       "PacketDebug": true
   },
   "Network": {
-    "AuthIp": "127.0.0.1",
-    "AuthPort": 4502,
-    "AuthServerIndex": 1,
-    "UploadIp": "127.0.0.1",
-    "UploadPort": 4616,
-    "Ip": "127.0.0.1",
-    "Port": 4515,
+    "Auth": {
+      "Ip": "127.0.0.1",
+      "Port": 4502,
+      "Index": 1
+    },
+    "Game": {
+      "Ip": "127.0.0.1",
+      "Port": 4515
+    },
+    "Upload": {
+      "UploadIp": "127.0.0.1",
+      "UploadPort": 4616
+    },
     "Backlog": 100,
-    "BufferSize": 32768,
     "MaxConnections": 2000,
+    "BufferSize": 32768,
     "CipherKey": "}h79q~B%al;k'y $E"
   },
   "Database": {
