+++ Network/Packets/Actions/AuthActions.cs
@@ -1,5 +1,4 @@
-﻿using Configuration;
-using Navislamia.Network.Enums;
+﻿using Navislamia.Network.Enums;
 using Network;
 using Navislamia.Notification;
 using System;
