+++ World/WorldModule.cs
@@ -1,6 +1,4 @@
-﻿using Configuration;
-using Navislamia.Notification;
-using System;
+﻿using Navislamia.Notification;
 
 namespace Navislamia.World
 {
